contextual_reasoning_ai/requirements.txt
# FastAPI API dependencies
fastapi==0.109.2
uvicorn[standard]==0.23.2
pydantic==2.5.1

# Database connectors
neo4j==6.0.0a1
psycopg2-binary==2.9.9
redis==6.4.0

# AI / ML
torch==2.2.2
transformers==4.34.0
numpy==1.26.2
scipy==1.11.1
scikit-learn==1.3.2

# Audio / Image
Pillow==10.1.0
librosa==0.10.1
soundfile==0.12.1
torchaudio==2.2.2

# GUI
tk==0.1.0  # Tkinter bindings for Python (should come with Python, included for container clarity)

# Optional utilities
python-dotenv==1.0.1
requests==2.32.3
